pyproject.toml
[tool.poetry]
name = "glider"
version = "0.1.0"
description = "The Bat Signal for your Space."
authors = ["Piotr Gaczkowski <doomhammerng@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
aiocron = "^1.6"
discord = "^1.7.3"
python-dotenv = "^0.19.0"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"